Moved health check functions to their own file
